Add blankline & Add dependency

+++ ports/netcpp/portfile.cmake
@@ -8,15 +8,19 @@ vcpkg_from_github(
     HEAD_REF release
 )
 
+if(!VCPKG_TARGET_IS_WINDOWS)
+    set(VCPKG_POLICY_EMPTY_PACKAGE enabled)
+    return()
+endif()
 
 vcpkg_cmake_configure(
     SOURCE_PATH "${SOURCE_PATH}"
 )
 
 vcpkg_cmake_install()
-
 vcpkg_cmake_config_fixup(CONFIG_PATH lib/cmake/netcpp PACKAGE_NAME netcpp)
-file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/debug/include")
 
+file(REMOVE_RECURSE "${CURRENT_PACKAGES_DIR}/debug/include")
 file(INSTALL "${SOURCE_PATH}/LICENSE" DESTINATION "${CURRENT_PACKAGES_DIR}/share/${PORT}" RENAME copyright)
+
 configure_file("${CMAKE_CURRENT_LIST_DIR}/usage" "${CURRENT_PACKAGES_DIR}/share/${PORT}/usage" COPYONLY)
